requirements.txt
selenium~=3.141.0
cassandra-driver~=3.25.0
requests~=2.27.1
cloudinary~=1.32.0
python-dotenv~=1.0.0